add svgs for navBar
